backend/requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
asyncpg>=0.29.0
greenlet>=3.0.0
aiosqlite>=0.19.0

# Queue
celery>=5.3.0
redis>=5.0.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Storage
boto3>=1.34.0
paramiko>=3.4.0

# HTTP/WebSocket
httpx>=0.26.0
websockets>=12.0

# Config/Utils
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Video/Audio
ffmpeg-python>=0.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
aiosqlite>=0.19.0

# Type checking
mypy>=1.8.0

# Linting
ruff>=0.1.0
